Move map const values to dedicaced file
